proxy/requirements.txt
bcrypt==4.1.2
blinker==1.7.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
paramiko==3.4.0
protobuf==4.21.12
pycparser==2.21
PyNaCl==1.5.0
sshtunnel==0.4.0
Werkzeug==3.0.1
